move api test to xgrid
